requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
confusable-homoglyphs==3.2.0
Django==4.1.1
django-registration==3.3
djangorestframework==3.13.1
pytz==2022.2.1
sqlparse==0.4.2